requirements.txt
discord.py >= 2.1.0
pyaml >= 6.0
requests >= 2.28.2